tariff_titan/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.1
pydantic>=2.7.4
pydantic-settings>=2.3.4
mistralai>=1.0.3
google-generativeai>=0.7.1
httpx>=0.23.0
python-multipart>=0.0.9
pandas>=2.2.2
openpyxl>=3.1.5
langchain-core>=0.2.10
langchain-community>=0.2.6
sentence-transformers>=3.0.1
langchain-huggingface>=0.1.2
pypdf>=4.2.0
